Refactor: Move setup wizard components to separate files
This commit refactors the setup wizard by extracting individual steps into their own Vue components. This improves code organization and maintainability.
