+++ src/views/setup/BasicSettingsStep.vue
