+++ src/views/setup/ConnectivityTest.vue
