+++ src/views/setup/DownloaderSettingsStep.vue
