+++ src/views/setup/MediaServerSettingsStep.vue
