+++ src/views/setup/PreferencesSettingsStep.vue
