+++ src/views/setup/StorageSettingsStep.vue
